fan: Scale fan speed between off_below and max_power
New relative_power fan setting will allow the scaling of fan speed requests
between  off_below and max_power. For example, if off_below is .3 and
max_power is .9 and a fan speed of 20% is requested, then the fan power
will be set to 42%).

With this setting enabled, once you have off_below and
max_power properly calibrated, it is much easier to request fan speeds
between 0% and 100% (or 0-255) via slicer for example and have the fan
behave as intended without having to be mindful of what speed the fan
will be under (ie not started) or over powered.